refactor: 역할별 패키지 분리
