+++ src/main/kotlin/racingcar/domain/car/CarInfo.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.car
 
 data class CarInfo(
     val name: String,
