+++ src/main/kotlin/racingcar/domain/car/Cars.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.car
 
 import racingcar.domain.movingstrategy.MovingStrategy
