+++ src/test/kotlin/racingcar/domain/car/PositionTest.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.car
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
